+++ backend/src/main/kotlin/de/vinz/openfls/domains/assistancePlans/services/AssistancePlanService.kt
@@ -11,7 +11,7 @@ import de.vinz.openfls.domains.clients.ClientService
 import de.vinz.openfls.domains.hourTypes.HourTypeService
 import de.vinz.openfls.domains.assistancePlans.dtos.ActualTargetValueDto
 import de.vinz.openfls.domains.assistancePlans.dtos.AssistancePlanEvalDto
-import de.vinz.openfls.domains.hourTypes.dtos.HourTypeDto
+import de.vinz.openfls.domains.hourTypes.HourTypeDto
 import de.vinz.openfls.domains.institutions.InstitutionService
 import de.vinz.openfls.domains.services.ServiceService
 import de.vinz.openfls.domains.sponsors.SponsorService
@@ -228,6 +228,15 @@ class AssistancePlanService(
         val end = LocalDate.of(year, month, 1).plusMonths(1).minusDays(1)
 
         return assistancePlanRepository.findProjectionByStartAndEnd(start, end)
+    }
+
+    fun getProjectionByYearMonthSponsorId(year: Int,
+                                          month: Int,
+                                          sponsorId: Long): List<AssistancePlanProjection> {
+        val start = LocalDate.of(year, month, 1)
+        val end = LocalDate.of(year, month, 1).plusMonths(1).minusDays(1)
+
+        return assistancePlanRepository.findProjectionBySponsorIdAndStartAndEnd(sponsorId, start, end)
     }
 
     fun getProjectionByYearMonthInstitutionId(year: Int,