+++ backend/src/main/kotlin/de/vinz/openfls/domains/categories/dtos/CategoryTemplateDto.kt
@@ -1,16 +1,23 @@
 package de.vinz.openfls.domains.categories.dtos
 
-import jakarta.validation.constraints.NotEmpty
+import de.vinz.openfls.domains.categories.entities.CategoryTemplate
 
-class CategoryTemplateDto {
-    var id: Long = 0
-
-    @field:NotEmpty
-    var title: String = ""
-
-    var description: String = ""
-
-    var withoutClient: Boolean = false
-
-    var categories: Array<CategoryDto> = emptyArray()
+data class CategoryTemplateDto(
+        var id: Long = 0,
+        var title: String = "",
+        var description: String = "",
+        var withoutClient: Boolean = false,
+        var categories: List<CategoryDto> = emptyList()
+) {
+    companion object {
+        fun from(categoryTemplate: CategoryTemplate): CategoryTemplateDto {
+            return CategoryTemplateDto(
+                    id = categoryTemplate.id,
+                    title = categoryTemplate.title,
+                    description = categoryTemplate.description,
+                    withoutClient = categoryTemplate.withoutClient,
+                    categories = categoryTemplate.categories.map { CategoryDto.from(it) }.sortedBy { it.title }
+            )
+        }
+    }
 }