+++ backend/src/main/kotlin/de/vinz/openfls/domains/contingents/dtos/ContingentDto.kt
@@ -1,5 +1,7 @@
 package de.vinz.openfls.domains.contingents.dtos
 
+import de.vinz.openfls.domains.contingents.Contingent
+import de.vinz.openfls.domains.contingents.projections.ContingentProjection
 import jakarta.validation.constraints.NotNull
 import java.time.LocalDate
 
@@ -19,4 +21,28 @@ class ContingentDto {
 
     @field:NotNull
     var institutionId: Long = 0
+
+    companion object {
+        fun from(entity: Contingent): ContingentDto {
+            val dto = ContingentDto()
+            dto.id = entity.id ?: 0
+            dto.start = entity.start
+            dto.end = entity.end
+            dto.weeklyServiceHours = entity.weeklyServiceHours
+            dto.employeeId = entity.employee?.id ?: 0
+            dto.institutionId = entity.institution?.id ?: 0
+            return dto
+        }
+
+        fun from(entity: ContingentProjection): ContingentDto {
+            val dto = ContingentDto()
+            dto.id = entity.id
+            dto.start = entity.start
+            dto.end = entity.end
+            dto.weeklyServiceHours = entity.weeklyServiceHours
+            dto.employeeId = entity.employee.id
+            dto.institutionId = entity.institution.id
+            return dto
+        }
+    }
 }