+++ backend/src/main/kotlin/de/vinz/openfls/domains/contingents/dtos/EmployeeContingentEvaluationDto.kt
@@ -1,11 +1,12 @@
 package de.vinz.openfls.domains.contingents.dtos
 
 data class EmployeeContingentEvaluationDto(
-        val employeeId: Long,
-        val lastname: String,
-        val firstname: String,
-        val contingentHours: List<Double>,
-        val executedHours: List<Double>,
-        val executedPercent: List<Double>,
-        val summedExecutedPercent: List<Double>,
-        val missingHours: List<Double>)
+    val employeeId: Long,
+    val lastname: String,
+    val firstname: String,
+    val contingentHours: List<Double>,
+    val executedHours: List<Double>,
+    val executedPercent: List<Double>,
+    val summedExecutedPercent: List<Double>,
+    val missingHours: List<Double>
+)