+++ backend/src/main/kotlin/de/vinz/openfls/domains/services/dtos/ServiceDto.kt
@@ -22,6 +22,8 @@ class ServiceDto {
 
     var unfinished: Boolean = false
 
+    var groupService: Boolean = false
+
     var minutes: Int = 0
 
     @field:NotNull