+++ frontend/src/app/pages/service-evaluation-overview/modals/overview-valuetype-info-modal/overview-value-type-info-modal.component.ts
@@ -11,7 +11,10 @@ export class OverviewValueTypeInfoModalComponent implements OnInit {
   readonly CLOSE_BUTTON_DESCRIPTION: string = "Schließen"
   readonly EXECUTED_TITLE: string = "geleistete Stunden"
   readonly EXECUTED_DESCRIPTION: string = "Geleistete Stunden umfassen die Summe aller Minuten der Einträge, " +
-    "die den jeweiligen Filtern entsprechen."
+    "die den jeweiligen Filtern entsprechen (darunter zählen auch die Gruppenangebote)."
+  readonly EXECUTED_GROUP_SERVICE_TITLE: string = "geleistete Stunden im Gruppenangebot"
+  readonly EXECUTED_GROUP_SERVICE_DESCRIPTION: string = "Geleistete Stunden umfassen die Summe aller Minuten der Einträge, " +
+    "die den jeweiligen Filtern entsprechen und als Gruppenangebote stattfanden."
   readonly APPROVED_TITLE: string = "genehmigte Stunden"
   readonly APPROVED_DESCRIPTION: string = "Genehmigte Stunden werden folgendermaßen berechnet. " +
     "Sofern ein Hilfeplan über eingetragene Stunden verfügt, so werden diese für die Berechnung herangezogen. " +